User controls and ICP changes.
Made some changes to ICP algorithm to attempt to make it work.
Made a bugfix to umeyama.js, because its math was wrong.
ICP algorithm probably works correctly now, but is has bad inputs.
ICP algorithm probably works best if one group is a subset of the other.
However, currently neither is a subset of the other because the LIDAR
scan contains surrounding buildings and the IFC schematic contains the
interior of the building.
Have some ideas to fix this, such as the ability to weight certain
points that are shared, and put low weight to unshared points.
That is added to umeyama.js, but probably not best plan.
Current plan is to have the user crop the LIDAR scan to isolate the
building. Then the LIDAR scan will be a subset of the IFC schematic.
Following this plan, I am now working on the user controls.

Added 3 control boxes automatically. One around the LIDAR scan, one
around the schematic, and one for the cropping box.
Added/modified control sidebar: sidebar from potree but with many things
 removed, and some custom options.
Most custom options don't work yet, but the selectgroup to pick a
controlbox to manipulate does work.
Changes to InputHandler: prevent user from switching controlbox via
clicking/delete key.
Changes to TransformationTool: remove yellow frame from controlbox
controls (there should only be a frame around the cropping box).