Bugfixes
Some pointcloud points would not be highlighted in blue and red, even when they are used by the algorithm.
Fixed this bug, but this new solution results in some points being highlighted twice. This may be visually indistinguishable however, and is probably preferable.

Also corrected README.txt instructions, where they asked for a file which was not needed and had no impact.